+++ app/containers/DetailContents.js
@@ -4,7 +4,6 @@ import {fetchSensorMobile, fetchSensor } from "../actions";
 import { getMobileSizeMax} from '../utils/getConfig';
 import type {Dispatch} from "../utils/flowtype";
 
-
 const mapStateToProps = (state, ownProps) => {
     return {
         sensorData: state.sensorDetail.datapoints,
@@ -14,14 +13,14 @@ const mapStateToProps = (state, ownProps) => {
 
 const mapDispatchToProps = (dispatch: Dispatch) => {
     return {
-        loadSensor: (id, name, use_season) => {
+        loadSensor: (id, name, use_season, binType, selectedStartDate, selectedEndDate) => {
             if(screen.width <= getMobileSizeMax()) {
                 dispatch(fetchSensorMobile(name))
             } else {
                 if(use_season) {
-                    dispatch(fetchSensor( name, 'semi'))
+                    dispatch(fetchSensor( name, 'season', selectedStartDate, selectedEndDate))
                 } else {
-                    dispatch(fetchSensor(name, 'day'))
+                    dispatch(fetchSensor(name, binType, selectedStartDate, selectedEndDate))
                 }
             }
         }